[UI] Phase 5.a: Split MesheryPatterns into focused siblings (#18660)
MesheryPatterns.tsx was the largest remaining giant in ui/ at ~1,679
raw lines. This change splits it mechanically into a handful of
co-located sibling files so the entry point drops below the 600-line
warn threshold, retiring its `legacyMaxLineOffenders` allowlist entry
and reducing `audit:size over_hard` from 4 to 3.

Behavior, the default export, the prop signature, the Redux/RTK state
shape, and the rendered DOM are unchanged. Each extracted file is a
verbatim lift of the original code wired back through a factory or a
prop-driven sub-component — no hex literals or MUI imports were
migrated (those stay allowlisted for a later Phase-2 pass).

New sibling files (all in components/designs/patterns/):
  - MesheryPatterns.styled.tsx        styled() definitions
  - MesheryPatterns.constants.tsx     ACTION_TYPES + small helpers
  - MesheryPatterns.columns.tsx       table columns + options builders
  - MesheryPatternsToolbar.tsx        header toolbar JSX
  - MesheryPatternsModals.tsx         ImportDesignModal + PublishModal
  - YAMLEditor.tsx                    in-place YAML edit dialog
  - design-lifecycle-handlers.tsx     deploy/undeploy/dryrun/validate
                                      modal openers
  - patterns-actions.ts               CRUD / publish / clone /
                                      download handler factory

eslint.config.js:
  - MesheryPatterns.tsx removed from legacyMaxLineOffenders
  - new files added to legacyLiteralColorOffenders and
    legacyInlineStyleOffenders where applicable

AUDIT size over_warn=11 over_hard=3

Refs #18660
Refs #18654
